requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreschema==0.0.4
Django==1.8.4
django-webpack-loader==0.2.4
djangorestframework==3.3.2
httpie==0.9.9
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
psycopg2==2.7.3.1
Pygments==2.2.0
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
